recommender-ai-service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-decouple==3.8
requests==2.31.0
django-cors-headers==4.3.1
numpy==1.26.2
scikit-learn==1.3.2
